# Runtime SSE4.2 dispatch