@indutny 2019-07-12
Hello!

Thanks for testing it. The results are very interesting!

It looks like what I was afraid of has manifested itself very clearly in benchmarks. The dynamic dispatch takes a lot of time to do. The apparently problem is that the check always runs.

JIT compilers usually solve this by hot-patching the call-sites to short-circuit them into either of supported modes of operation (SSE or no-SSE), but this is very problematic for C language.

I'm not sure if I have any constructive ideas or suggestions on how to move forward with this experiment. Sorry!