@zbjornson 2019-07-12
I take it there's no Node subclass or something that always emits a function instead of emitting source that's inlined into another function? This can be made ~zero-cost if there is.

Do you have plans for how to make the `pcmpestri` change benefit Node.js generally, or will it be limited to users who build their own Node.js with custom `-m` flags?