@zbjornson 2019-07-12
(It looks like helpers/match-sequence.ts emits a standalone function, for example. Making a helpers/table-lookup.ts could work if that pattern is reusable.)